Rate tracker to track by second instead of ms
